:fire: remove ModelService and rename
