+++ modelci/app/experimental/endpoints/cv_tuner.py
@@ -13,7 +13,7 @@ from modelci.experimental.model.model_structure import Structure, Operation
 from modelci.hub.registrar import register_model
 from modelci.hub.cache_manager import get_remote_model_weight
 from modelci.hub.utils import generate_path_plain
-from modelci.persistence.service_ import get_by_id, get_models
+from modelci.persistence.service import get_by_id, get_models
 from modelci.types.models import MLModel, Engine, IOShape, ModelStatus
 from modelci.types.type_conversion import model_data_type_to_torch, type_to_data_type
 from modelci.utils.exceptions import ModelStructureError
