+++ modelci/hub/registrar.py
@@ -28,7 +28,7 @@ from modelci.hub.client.torch_client import CVTorchClient
 from modelci.hub.client.trt_client import CVTRTClient
 from modelci.hub.converter import converter
 from modelci.hub.utils import parse_path_plain
-from modelci.persistence.service_ import save
+from modelci.persistence.service import save
 from modelci.types.models import MLModelFromYaml, MLModel
 from urllib.request import urlopen, Request
 from tqdm.auto import tqdm
